Minor Framework Update
- the concept of a SceneID was dropped and replaced with a
proper definition of a session ID:
"A SessionID is the unique identifier for a single instance of a user of the framework. For example, each execution of a software component (for example: an instance of a game) connecting to the MOSIM framework should use a unique identifier to connect to the framework. This means,  when stopping and restarting a game it should connect with a different session ID on the second run. Thus, the SessionID describes a single environment, in which the agents are simulated. In case, a new game map is loaded, push scene must be utilized to update the sessions scene in the different adapters. "

- The avatarID was still missing in the checkpoint methods inside
the adapter.